refactor: Arreglar nombre de carpetas del proyecto
