+++ src/components/sCard/Card.css
