+++ src/components/sHeader/Header.css
@@ -5,8 +5,11 @@
   display: flex;
   flex-direction: row;
   align-items: center;
-
+  position: fixed;
   justify-content: center;
+  top: 0;
+  width: 100vw;
+  z-index: 10;
 }
 
 .header__container {
