+++ src/components/sLogin/Login.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Form from "../Form/Form";
+import Form from "../sForm/Form";
 import "./Login.css";
 import { Link } from "react-router-dom";
 
@@ -7,7 +7,7 @@ function Login() {
   return (
     <section className="login">
       <div className="login__container">
-        <Form title="Iniciar Sesión" nameBtn="Iniciar sesión">
+        <Form title="Iniciar Sesión">
           <p className="login_paragraph">
             ¿Aún no eres miembro?{" "}
             <Link to="/signup" className="login__link">
