+++ src/components/sMain/Main.jsx
@@ -1,11 +1,13 @@
 import React from "react";
 import "./Main.css";
-import Present from "../Present/Present";
+import Present from "../sPresent/Present";
+import Portfolios from "../sPortfolios/Portfolios";
 
 function Main() {
   return (
     <main className="main">
       <Present />
+      <Portfolios />
     </main>
   );
 }
