+++ src/components/sPortfolios/portfolios.css
@@ -1,0 +1,11 @@
+.portfolios {
+  border: solid 1px red;
+}
+
+.portfolios__content {
+  margin-top: 5%;
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
+  grid-template-rows: repeat(auto-fill, minmax(200px, 1fr));
+  gap: 30px;
+}
