+++ src/components/sPresent/Present.css
@@ -15,7 +15,7 @@
   width: 750px;
   height: 400px;
   border-radius: 20px;
-  margin: 5% auto 0;
+  margin: 20% auto 0;
 }
 
 .present__name-user {
@@ -25,4 +25,21 @@
   white-space: nowrap;
 
   overflow: hidden;
+}
+
+.present__btn {
+  border: solid 2px #844eb8;
+  background-color: white;
+  font-size: 18px;
+  margin: 8% auto;
+  border-radius: 20px;
+  width: 200px;
+  height: 60px;
+  cursor: pointer;
+  transition: color 0.2s ease-in-out, 0.2s ease-in-out;
+}
+
+.present__btn:hover {
+  color: #0255e5;
+  font-size: 20px;
 }
