+++ src/components/sRegister/Register.jsx
@@ -1,13 +1,13 @@
 import React from "react";
 import "./Register.css";
-import Form from "../Form/Form";
+import Form from "../sForm/Form";
 import { Link } from "react-router-dom";
 
 function Register() {
   return (
     <section className="register">
       <div className="register__container">
-        <Form title="Inscribirse" nameBtn="Regístrate">
+        <Form title="Inscribirse">
           <p className="register_paragraph">
             ¿Ya eres miembro? Inicia{" "}
             <Link to="/signin" className="register__link">
